// src/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const Role = {
  ADMIN: 'ADMIN',
  MEMBER: 'MEMBER'
} as const

export type Role = (typeof Role)[keyof typeof Role]


export const TalentType = {
  INHOUSE: 'INHOUSE',
  PARTNER: 'PARTNER'
} as const

export type TalentType = (typeof TalentType)[keyof typeof TalentType]


export const DataSource = {
  REGISTER: 'REGISTER',
  EMAIL: 'EMAIL'
} as const

export type DataSource = (typeof DataSource)[keyof typeof DataSource]


export const TalentStatus = {
  NONE: 'NONE',
  PROPOSING: 'PROPOSING',
  ACTIVE: 'ACTIVE',
  CLOSED: 'CLOSED'
} as const

export type TalentStatus = (typeof TalentStatus)[keyof typeof TalentStatus]


export const ProjectStatus = {
  OPEN: 'OPEN',
  PROPOSING: 'PROPOSING',
  DECIDED: 'DECIDED',
  CLOSED: 'CLOSED'
} as const

export type ProjectStatus = (typeof ProjectStatus)[keyof typeof ProjectStatus]


export const Gender = {
  MALE: 'MALE',
  FEMALE: 'FEMALE',
  OTHER: 'OTHER'
} as const

export type Gender = (typeof Gender)[keyof typeof Gender]


export const EmploymentType = {
  EMPLOYEE: 'EMPLOYEE',
  FREELANCE: 'FREELANCE'
} as const

export type EmploymentType = (typeof EmploymentType)[keyof typeof EmploymentType]


export const Nationality = {
  JAPAN: 'JAPAN',
  OTHER: 'OTHER'
} as const

export type Nationality = (typeof Nationality)[keyof typeof Nationality]


export const LanguageLevel = {
  NATIVE: 'NATIVE',
  BUSINESS: 'BUSINESS',
  DAILY: 'DAILY',
  NONE: 'NONE'
} as const

export type LanguageLevel = (typeof LanguageLevel)[keyof typeof LanguageLevel]


export const RemotePreference = {
  FULL_REMOTE: 'FULL_REMOTE',
  MOSTLY_REMOTE: 'MOSTLY_REMOTE',
  HYBRID: 'HYBRID',
  OFFICE_1: 'OFFICE_1',
  OFFICE_2: 'OFFICE_2',
  OFFICE_3: 'OFFICE_3',
  OFFICE_4: 'OFFICE_4',
  ONSITE: 'ONSITE'
} as const

export type RemotePreference = (typeof RemotePreference)[keyof typeof RemotePreference]